requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.1.7
django-debug-toolbar==1.11
django-shortuuidfield==0.1.3
djangorestframework==3.9.1
idna==2.8
mysqlclient==1.3.14
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.3
pyOpenSSL==18.0.0
PySocks==1.6.8
python-memcached==1.59
pytz==2018.9
qiniu==7.2.2
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
win-inet-pton==1.0.1
wincertstore==0.2
